implement Memory Task 1
