+++ fe/fe-core/src/main/java/org/apache/doris/scheduler/manager/MemoryTaskManager.java
@@ -1,0 +1,57 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package org.apache.doris.scheduler.manager;
+
+import org.apache.doris.scheduler.disruptor.TimerTaskDisruptor;
+import org.apache.doris.scheduler.executor.MemoryTaskExecutor;
+
+import lombok.Setter;
+
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+public class MemoryTaskManager {
+    /**
+     * key: taskId
+     * value: memory task executor of this task
+     * it's used to star task
+     */
+    private final ConcurrentHashMap<Long, MemoryTaskExecutor> taskExecutorMap = new ConcurrentHashMap<>(128);
+
+    /**
+     * Producer and Consumer model
+     * disruptor is used to handle task
+     * disruptor will start a thread pool to handle task
+     */
+    @Setter
+    private TimerTaskDisruptor disruptor;
+
+    public MemoryTaskManager() {
+    }
+
+    public MemoryTaskExecutor getMemoryTaskExecutor(Long taskId) {
+        return taskExecutorMap.get(taskId);
+    }
+
+    public Long registerMemoryTask(MemoryTaskExecutor executor) {
+        Long taskId = UUID.randomUUID().getMostSignificantBits();
+        taskExecutorMap.put(taskId, executor);
+        disruptor.tryPublishTask(taskId);
+        return taskId;
+    }
+}
